+++ server/schema/types/Question.js
@@ -16,12 +16,34 @@ const typeDefs = `
     test: String!,
     solution: String!
   }
+  type QuestionHeader {
+    _id: ID!
+    challenge: String!
+    languages: [String]!
+  }
+  extend type Query {
+    getQuestions(filter: String!): [QuestionHeader]
+  }
   extend type Mutation {
     addQuestion: String!
   }
 `;
 
 const resolvers = {
+  Query: {
+    async getQuestions(_, { filter }, ___) {
+
+      const questions = await Question.find({
+        challenge: new RegExp(filter, 'i')
+      });
+
+      return questions.map(q => ({
+        _id: q._id.toString(),
+        challenge: q.challenge,
+        languages: q.functionNames.map(fn => fn.language)
+      }));
+    }
+  },
   Mutation: {
     addQuestion(_, __, ___) {
       const fizzBuzzReference = function (n) {